+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * Don't use O_TMPFILE for memory map file, as it may not be available (#524)
 * Check and update movie length during last frame instead of one frame later
   and mark it as modified
+* Enforce No Sleep on Unity games
 
 ### Fixed